+++ src/app/battery-level-widget/river-2-pro/river-2-pro.ts
@@ -1,20 +1,19 @@
 import { Component, inject, input, OnDestroy, OnInit, signal } from '@angular/core';
 import { UnlistenFn } from '@tauri-apps/api/event';
-import { Card } from 'primeng/card';
+import { Menu } from '@tauri-apps/api/menu';
+import { getCurrentWindow } from '@tauri-apps/api/window';
+import { Button } from 'primeng/button';
 import { ProgressBar } from 'primeng/progressbar';
-import { QuotaPayload } from './payloads/quota-payload';
-import { River2ProCommunicationService } from './river-2-pro-communication-service';
+import { QuotaPayload } from '../../services/devices/quota-payload';
+import { River2ProCommunicationService } from '../../services/devices/river-2-pro-communication-service';
 
 @Component({
   selector: 'edc-river-2-pro',
   templateUrl: './river-2-pro.html',
   styleUrl: './river-2-pro.scss',
   imports: [
-    Card,
     ProgressBar,
-  ],
-  providers: [
-    River2ProCommunicationService,
+    Button,
   ],
 })
 export class River2Pro implements OnInit, OnDestroy {
@@ -41,11 +40,30 @@ export class River2Pro implements OnInit, OnDestroy {
   }
 
 
+  async showMenu() {
+    const menu = await Menu.new({
+      items: [
+        {
+          id: 'close',
+          text: 'Close',
+          action: this.onMenuCloseRequested.bind(this),
+        },
+      ]
+    });
+    await menu.popup();
+  }
+
+
   private onPowerDeliveryPayloadReceived(payload: QuotaPayload<'pdStatus'>) {
     this.batteryLevel.set(payload.params.soc);
     const totalRemainMinutes = payload.params.remainTime;
     const remainHours = Math.floor(totalRemainMinutes / 60).toString().padStart(2, '0');
     const remainMinutes = Math.floor(totalRemainMinutes % 60).toString().padStart(2, '0');
     this.remainTime.set(`${remainHours}:${remainMinutes}`);
+  }
+
+
+  private async onMenuCloseRequested() {
+    await getCurrentWindow().close();
   }
 }